Final changes for today from Armand.
Updated description with names of team members, date the project is due, descroption of the project. Refactored code to be more readable and adhere more to .NET coding standards. Still a bunch of code I left in there that I may use. I'll leave refactoring all of that for last.